backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
channels==4.1.0
django-cors-headers==4.4.0
pymongo==4.10.1
Pillow==10.4.0
